working FASTA support

+++ plasmidtron/FastqReadNames.py
@@ -17,6 +17,10 @@ class FastqReadNames:
 	
 	def extract_readnames_from_fastq(self):
 		self.logger.warning("Extracting read names from FASTQ file")
+		if not os.path.exists(self.fastq_file):
+			self.logger.error('Cannot read the FASTQ file %s', self.fastq_file)
+			raise
+		
 		with open(self.fastq_file, "r") as fastq_file_input, open(self.output_readnames_file, "w") as readnames_output:
 			for record in SeqIO.parse(fastq_file_input, "fastq"):
 				readnames_output.write(record.id + '\n')
